# backend/docs/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "api.K8sObject": {
        "additionalProperties": true,
        "type": "object"
      },
      "api.Response": {
        "properties": {
          "code": {
            "example": 200,
            "type": "integer"
          },
          "data": {},
          "message": {
            "example": "Success",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "输入格式: Bearer {session_id}，session_id 从登录接口获取 (Swagger 暂不支持 Cookie 自动发送，请手动输入)",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@example.com",
      "name": "API Support",
      "url": "http://www.example.com/support"
    },
    "description": "DevOps 运维平台接口文档，登录后在 Authorize 中输入 Bearer {session_id} 即可调用接口",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "DevOps 运维平台 API",
    "version": "1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/k8s/configmap/create": {
      "post": {
        "description": "创建一个新的 ConfigMap",
        "parameters": [
          {
            "description": "集群ID（可选，未传则使用默认集群）",
            "in": "query",
            "name": "clusterId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "命名空间",
            "in": "query",
            "name": "namespace",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api.K8sObject"
              }
            }
          },
          "description": "ConfigMap 对象",
          "required": true,
          "x-originalParamName": "configmap"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api.Response"
                }
              }
            },
            "description": "成功"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "创建 ConfigMap",
        "tags": [
          "K8s资源管理"
        ]
      }
    },
    "/k8s/configmap/delete": {
      "post": {
        "description": "删除指定的 ConfigMap",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "description": "参数: {clusterId, namespace, name}",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api.Response"
                }
              }
            },
            "description": "成功"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "删除 ConfigMap",
        "tags": [
          "K8s资源管理"
        ]
      }
    },
    "/k8s/configmap/detail": {
      "get": {
        "description": "获取指定 ConfigMap 的详细信息",
        "parameters": [
          {
            "description": "集群ID（可选，未传则使用默认集群）",
            "in": "query",
            "name": "clusterId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "命名空间",
            "in": "query",
            "name": "namespace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "资源名称",
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api.Response"
                }
              }
            },
            "description": "成功"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "获取 ConfigMap 详情",
        "tags": [
          "K8s资源管理"
        ]
      }
    },
    "/k8s/configmap/list": {
      "get": {
        "description": "获取 ConfigMap 列表，namespace 为空时获取所有命名空间",
        "parameters": [
          {
            "description": "集群ID（可选，未传则使用默认集群）",
            "in": "query",
            "name": "clusterId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "命名空间（为空时查询所有命名空间）",
            "in": "query",
            "name": "namespace",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "页码",
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "description": "每页数量",
            "in": "query",
            "name": "pageSize",
            "schema": {
              "default": 10,
              "type": "integer"
            }
          },
          {
            "description": "关键字搜索（匹配名称）",
            "in": "query",
            "name": "keyword",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api.Response"
                }
              }
            },
            "description": "成功"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "获取 ConfigMap 列表",
        "tags": [
          "K8s资源管理"
        ]
      }
    },
    "/k8s/configmap/update": {
      "post": {
        "description": "更新现有的 ConfigMap",
        "parameters": [
          {
            "description": "集群ID（可选，未传则使用默认集群）",
            "in": "query",
            "name": "clusterId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "命名空间",
            "in": "query",
            "name": "namespace",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api.K8sObject"
              }
            }
          },
          "description": "ConfigMap 对象",
          "required": true,
          "x-originalParamName": "configmap"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api.Response"
                }
              }
            },
            "description": "成功"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "更新 ConfigMap",
        "tags": [
          "K8s资源管理"
        ]
      }
    },
...
    "/k8s/namespaces/list": {
      "get": {
        "description": "获取指定集群下的所有 Namespace 列表",
        "parameters": [
          {
            "description": "集群ID（可选，未传则使用默认集群）",
            "in": "query",
            "name": "clusterId",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api.Response"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api.Response"
                }
              }
            },
            "description": "参数错误"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api.Response"
                }
              }
            },
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "获取 Namespace 列表",
        "tags": [
          "K8s资源管理"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1"
    },
    {
      "url": "https://localhost:8000/api/v1"
    }
  ]
}